Improve the OverlayPanel table integration demo
Move buttons to the right side of the table to show how
the triangles are automatically shifted to the right.
Use fixed sizes for the images because otherwise the size
of the panel can only be determined when the image has been loaded,
which is too late and can lead to misplacement of the panel.